requirements.txt
fastapi
uvicorn
pokebase
python-telegram-bot
pydantic
pydantic-settings
python-dotenv
